+++ src/produtos/produtos.service.ts
@@ -8,31 +8,60 @@ export class ProdutosService {
   constructor(private readonly prisma: PrismaService) {}
 
   async create(createProdutoDto: CreateProdutoDto) {
-    const produto = await this.prisma.produto.create({
+    const produto = await this.prisma.produto.findUnique({
+      where: {sku: createProdutoDto.sku},
+    });
+
+    if (produto) {
+      throw new Error('Produto já cadastrado');
+    }
+
+    return await this.prisma.produto.create({
       data: createProdutoDto,
     });
-    return produto;
   }
 
   async findAll() {
-    return await this.prisma.produto.findMany();
+    const produtos = await this.prisma.produto.findMany();
+    return produtos;
   }
 
   async findOne(id: number) {
-    return await this.prisma.produto.findUnique({
+   const produto = await this.prisma.produto.findUnique({
       where: { id },
     });
+
+    if (!produto) {
+      throw new Error('Produto não encontrado');
+    }
+    return produto;
   }
 
 
   async update(id: number, updateProdutoDto: UpdateProdutoDto) {
-    return await this.prisma.produto.update ({
+    const produto = await this.prisma.produto.findUnique({
+      where: { id },
+    });
+
+    if (!produto) {
+      throw new Error('Produto não encontrado');
+    }
+
+    return await this.prisma.produto.update({
       where: { id },
       data: updateProdutoDto,
     });
   }
 
   async remove(id: number) {
+    const produto = await this.prisma.produto.findUnique({
+      where: { id },
+    });
+
+    if (!produto) {
+      throw new Error('Produto não encontrado');
+    }
+
     return await this.prisma.produto.delete({
       where: { id },
     });
