+++ src/vendedores/vendedores.service.ts
@@ -7,23 +7,47 @@ import { PrismaService } from '../prisma/prisma.service';
 @Injectable()
 export class VendedoresService {
   constructor(private readonly prisma: PrismaService) {}
+
   async create(createVendedoreDto: CreateVendedoreDto) {
+   const vendedor = await this.prisma.vendedor.findUnique({
+      where: { email: createVendedoreDto.email },
+    });
+
+    if (vendedor) {
+      throw new Error('Vendedor já cadastrado');
+    }
+
     return await this.prisma.vendedor.create({
       data: createVendedoreDto,
     });
   }
 
   async findAll() {
-    return await this.prisma.vendedor.findMany();
+    const vendedores = await this.prisma.vendedor.findMany();
+    return vendedores;
   }
 
   async findOne(id: number) {
-    return await this.prisma.vendedor.findUnique({
+    const vendedor = await this.prisma.vendedor.findUnique({
       where: { id },
-    }); 
+    });
+
+    if (!vendedor) {
+      throw new Error('Vendedor não encontrado');
+    }
+
+    return vendedor;
   }
 
   async update(id: number, updateVendedoreDto: UpdateVendedoreDto) {
+    const vendedor = await this.prisma.vendedor.findUnique({
+      where: { id },
+    });
+
+    if (!vendedor) {
+      throw new Error('Vendedor não encontrado');
+    }
+
     return await this.prisma.vendedor.update({
       where: { id },
       data: updateVendedoreDto,
@@ -31,6 +55,14 @@ export class VendedoresService {
   }
 
   async remove(id: number) {
+    const vendedor = await this.prisma.vendedor.findUnique({
+      where: { id },
+    });
+
+    if (!vendedor) {
+      throw new Error('Vendedor não encontrado');
+    }
+
     return await this.prisma.vendedor.delete({
       where: { id },
     });
